move RenderTilesClient to tile sub-package
