+++ render-ws-java-client/src/main/java/org/janelia/render/client/tile/RenderTilesClient.java
@@ -1,4 +1,4 @@
-package org.janelia.render.client;
+package org.janelia.render.client.tile;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParametersDelegate;
@@ -47,6 +47,8 @@ import org.janelia.alignment.spec.stack.StackMetaData;
 import org.janelia.alignment.util.FileUtil;
 import org.janelia.alignment.util.ImageProcessorCache;
 import org.janelia.alignment.util.RenderWebServiceUrls;
+import org.janelia.render.client.ClientRunner;
+import org.janelia.render.client.RenderDataClient;
 import org.janelia.render.client.parameter.CommandLineParameters;
 import org.janelia.render.client.parameter.RenderWebServiceParameters;
 import org.janelia.saalfeldlab.googlecloud.GoogleCloudStorageURI;
